Wrote the parse function cpp file
The parse function is separated into a separate cpp file so that it can
be compiled and once compiled it need never be editted. This allows it
to be needful to only pay the cost of compilation *once*. This is good
because of the high cost associated with compiling heavily templated
code like Spirit::qi uses.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>