Many refactors and additions
Added a number of functions to separate and organize the drawing to the
screen of the various menus and information. Also added a set of placeholder
command execution functions. These functions contain the skeletons for
control and UI mode switching.

I am feeling a bit bad about this class though as it almost seems to be
becomming a bit of a monster with the large number of private variables and
functions that it has running. However, thus far I do not see a simple way
of breaking the beast up.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>