Improvements to everything
Partially implemented some of the command functions. Also added some
functions for handling key presses in getting run numbers and run titles.

Also swapped around the order of event loops. Now we draw, check for input,
then sleep.

Finally, put the mechanism to clear persistent messages when calling
commands, above the invocation of the command so that the command can set a
message and not have it instantly cleared.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>