Fixed const correctness
Using the mutex to block multiple threads from the runcommand function had
the side effect of making that function non-const. Due to this I had to go
through and make everything that called that function as well as that
function non const. Didn't think about that when I added the mutex... oops.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>