Added logging capabilities with boost::log
Was having some trouble with incorrect interpretation of the values returned
from snmp commands that gave status in "bits" format. To track this I needed
output, but the output needed to be thread safe and not interfere with the
ncurses stuff. To accomplish this I looked into the boost::log facility.
The changes in this log represent what is needed to make a log that is
globally accessible to all threads. All one needs to do is include
"Utility/OrchidLogger.h" and go from there.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>