Fixed a bug in the status string gen
On an unrelated note, there is a *rare* race condition in the UIThread
waiting for the MPOD to ramp down. It will only exist for as long as we are
spoofing the return values from the MPOD instead of getting the actual ones.
It occurs as follows. The UIThread signals the MpodController that it should
shut-down. In real life the Mpod controller dutifully starts ramping down
the voltages and all is well. In this spoofed version of real life the Mpod
controller does show things as switched off, but then if before the UIThread
sees that, the slow controlls thread does a measurement of the voltage it
will show that the channels are on still and that causes the UI thread to
wait forever to shut down, this cannot ever happen in real life and so I am
ignoring it.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>