Updated FileOutputThreadController class
I updated the class with a way to update run titles and numbers. I wound up
not needing to guard the string and int by using an atomic signalling
system. By using the atomic state variable to tell the file output thread
that it has new parameters I use that as a "guard" around the data. The
thread that changes the data (IE the UI thread) does not change the state to
"NewRunParams" until after the data is changed, therefor preventing the
FileThread from ever checking those parameters until the UI thread is
*definitely* done changing them.

To prevent the UI thread from changing them again while the file thread has
not absorbed the first changed, we check the current state before doing any
modifications. If the current state is set to "NewRunParamas" then we do not
change the state and instead return false.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>