New Functions and Improved Efficiency
Added funtions that the File output thread uses to get the new run number
and run title.

Also added an atomic boolean that is used to control if we call notify on
the wait condition. Since calls to waitCondition.notify*() can be expensive,
we make it such that we do not call it when nothing is waiting, if the
thread is not waiting, it will notice state changes on its own.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>