Fully implemented a queue pair system
This pair is used to move data from a producer to a consumer and then to
move empty data containers from consumer back to producer. This is
accomplished by using two queues, first a consumer queue takes full objects
from the producer to the consumer. Then a producer queue takes empty objects
from the consumer back to the consumer.

For short hand, a producer thread pushes onto the consumer queue and pops
from the producer queue. A consumer thread pushes onto the producer queue
and pops from the consumer queue.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>