Added this shell class
This tiny class handles the fact that the thread class is not copyable.
This class is and it holds a pointer to the class that is not. Therefor,
this class, which is copyable is what is actually called by the thread, but
all it does is forward that call to the main callable

Signed-off-by: James Till Matta <jamesmatta@gmail.com>