Now have to enter run info before starting
Modified the async file output class to handle not being instantly
initialized with a file name to write to. Also modified the file output
thread to handle that properly. Also modified the file thread data structure
to allow the transmission of 'level' of initialization status.

Finally the UI thread has been modified such that unless a run is fully
initialized, it will not start data taking.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>