More appropriate name and a new class
The RawBufferTriple class is a template class container that contains 3
things. First, an unsigned int* that contains the buffer. Second, an
unsigned int that contains the number of cells filled in the buffer. Third,
a type specified by the template parameter that contains additional
information about the data source needed for processing. Like module number.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>