Changed data directory and interrupt level
I am noticing that the busy light on the digitizer is coming on periodically,
to fight this I am changing the setting to 1 board aggregate raises an interrupt.

I then also changed the data location so as not to confuse the new setting data
with old data. I will run new calibrations with the new data.

Signed-Off-By: James Till Matta <jamesmatta@gmail.com>