Fixed a weird data race and added a sleep
The data race, which in normal usage should be super uncommon, could have
screwed things up if a really odd usage condition occurred. Here I have
fixed it. The second thing that I have done is to add a sleep function.
This allows the slow controls thread to sleep using the wait condition,
allowing it to be responsive to state changes, now instead of waiting for it
to wake up from it using this_thread::sleep(duration) the change of state
hits the wait condition that it is using to sleep in the first place causing
it to wake up immediately.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>