Fixed some semi-benign threading issues
The thread controllers had some race conditions that did not rear their
heads in normal human usage, but broke down when hammered by automatic code.
Additionally, the idiotic practice I was making of passing more than one
mutex to a wait condition was introducint the possibility of things breaking
badly on wakeup, which they didn't seem to do unless the system got into a
wierd state. The weird state only happened recently, I am wondering if it is
some sort of system update that brought a library into posix compliance

All of this was fixed in the towardsSecant branch, but that hadn't been
propogated to the master branch yet

Signed-off-by: James Till Matta <jamesmatta@gmail.com>