Replaced throws with aborts.
I use exceptions as a way to kill execution. I used them because they allow
me to insert an additional message into the exception so that what happened
can be seen on the terminal. However, I found out that the coredump from a
program killed by an exception is annoying. While the backtrace is
preserved, the local variables for the backtrace do not seem to exist. My
suspicion is that this is because the stack unwinding procedure destructed
those local variables prior to the abort. Anyways, in an attempt to prevent
that I am using std::abort now. The other possibility is that the locals got
optimized out, in which case this will do nothing

Signed-off-by: James Till Matta <jamesmatta@gmail.com>